+++ dotnet/samples/KernelSyntaxExamples/Example13_ConversationSummarySkill.cs
@@ -4,9 +4,9 @@ using System;
 using System.Collections.Generic;
 using System.Threading.Tasks;
 using Microsoft.SemanticKernel;
-using Microsoft.SemanticKernel.CoreSkills;
 using Microsoft.SemanticKernel.Orchestration;
 using Microsoft.SemanticKernel.SkillDefinition;
+using Microsoft.SemanticKernel.Skills.Core;
 using RepoUtils;
 
 // ReSharper disable once InconsistentNaming
