+++ dotnet/samples/KernelSyntaxExamples/Example15_MemorySkill.cs
@@ -3,9 +3,9 @@
 using System;
 using System.Threading.Tasks;
 using Microsoft.SemanticKernel;
-using Microsoft.SemanticKernel.CoreSkills;
 using Microsoft.SemanticKernel.Memory;
 using Microsoft.SemanticKernel.SkillDefinition;
+using Microsoft.SemanticKernel.Skills.Core;
 using RepoUtils;
 
 // ReSharper disable once InconsistentNaming
