+++ dotnet/samples/KernelSyntaxExamples/Example30_ChatWithPrompts.cs
@@ -5,7 +5,7 @@ using System.Globalization;
 using System.Threading.Tasks;
 using Microsoft.SemanticKernel;
 using Microsoft.SemanticKernel.AI.ChatCompletion;
-using Microsoft.SemanticKernel.CoreSkills;
+using Microsoft.SemanticKernel.Skills.Core;
 using Microsoft.SemanticKernel.TemplateEngine;
 using RepoUtils;
 using Resources;
