+++ dotnet/src/Skills/Skills.Core/SemanticFunctionConstants.cs
@@ -1,6 +1,6 @@
 ﻿// Copyright (c) Microsoft. All rights reserved.
 
-namespace Microsoft.SemanticKernel.CoreSkills;
+namespace Microsoft.SemanticKernel.Skills.Core;
 
 internal static class SemanticFunctionConstants
 {
