+++ dotnet/src/Skills/Skills.Core/TimeSkill.cs
@@ -4,7 +4,7 @@ using System;
 using System.ComponentModel;
 using Microsoft.SemanticKernel.SkillDefinition;
 
-namespace Microsoft.SemanticKernel.CoreSkills;
+namespace Microsoft.SemanticKernel.Skills.Core;
 
 /// <summary>
 /// TimeSkill provides a set of functions to get the current time and date.
