+++ dotnet/src/Skills/Skills.UnitTests/Core/MathSkillTests.cs
@@ -2,10 +2,11 @@
 
 using System.Threading.Tasks;
 using Microsoft.SemanticKernel;
-using Microsoft.SemanticKernel.CoreSkills;
+using Microsoft.SemanticKernel.Skills.Core;
+using SemanticKernel.UnitTests;
 using Xunit;
 
-namespace SemanticKernel.UnitTests.CoreSkills;
+namespace SemanticKernel.Skills.UnitTests.Core;
 
 public class MathSkillTests
 {
