+++ dotnet/src/Skills/Skills.UnitTests/Skills.UnitTests.csproj
@@ -42,6 +42,7 @@
 
   <ItemGroup>
     <ProjectReference Include="..\..\SemanticKernel\SemanticKernel.csproj" />
+    <ProjectReference Include="..\Skills.Core\Skills.Core.csproj" />
     <ProjectReference Include="..\Skills.Document\Skills.Document.csproj" />
     <ProjectReference Include="..\Skills.Grpc\Skills.Grpc.csproj" />
     <ProjectReference Include="..\Skills.MsGraph\Skills.MsGraph.csproj" />
